// src/services/central-api/generated/model/ctl-mst-color.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PL Cental API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CtlMstColor
 */
export interface CtlMstColor {
    /**
     * 
     * @type {string}
     * @memberof CtlMstColor
     */
    'code'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstColor
     */
    'colorName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstColor
     */
    'createTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstColor
     */
    'createBy'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof CtlMstColor
     */
    'isDelete'?: boolean | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstColor
     */
    'updateTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstColor
     */
    'updateBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstColor
     */
    'deleteTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstColor
     */
    'deleteBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstColor
     */
    'tenantCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstColor
     */
    'branchCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstColor
     */
    'status'?: string | null;
}